JAMES-1869: mailbox functions should limit the size of a mailbox name

+++ protocols/imap/src/main/java/org/apache/james/imap/decode/parser/DeleteCommandParser.java
@@ -21,10 +21,12 @@ package org.apache.james.imap.decode.parser;
 import org.apache.james.imap.api.ImapCommand;
 import org.apache.james.imap.api.ImapConstants;
 import org.apache.james.imap.api.ImapMessage;
+import org.apache.james.imap.api.display.HumanReadableText;
 import org.apache.james.imap.api.process.ImapSession;
 import org.apache.james.imap.decode.ImapRequestLineReader;
 import org.apache.james.imap.decode.base.AbstractImapCommandParser;
 import org.apache.james.imap.message.request.DeleteRequest;
+import org.apache.james.mailbox.model.MailboxConstants;
 import org.apache.james.protocols.imap.DecodingException;
 
 /**
@@ -44,6 +46,9 @@ public class DeleteCommandParser extends AbstractImapCommandParser {
      */
     protected ImapMessage decode(ImapCommand command, ImapRequestLineReader request, String tag, ImapSession session) throws DecodingException {
         String mailboxName = request.mailbox();
+        if (mailboxName.length() > MailboxConstants.DEFAULT_LIMIT_MAILBOX_NAME_SIZE) {
+            throw new DecodingException(HumanReadableText.ILLEGAL_ARGUMENTS, "The mailbox name is over limit: " + mailboxName);
+        }
         request.eol();
 
         return new DeleteRequest(command, mailboxName, tag);
